feat : ticket 관련 쿼리 수정
