+++ src/apis/UserInfoApi.tsx
@@ -4,9 +4,7 @@ import { IUserInfoRes } from './type/userInfo';
 const UserInfoApi = {
   // 유저 정보 가져오기
   getUserInfo: async () => {
-    const { data }: { data: IUserInfoRes } = await axiosPrivate.get(
-      '/v1/users',
-    );
+    const { data } = await axiosPrivate.get('/v1/users');
     return data;
   },
 };
